+++ lib/uglifier.rb
@@ -61,7 +61,9 @@ class Uglifier
       :evaluate => true, # Attempt to evaluate constant expressions
       :booleans => true, # Various optimizations to boolean contexts
       :loops => true, # Optimize loops when condition can be statically determined
-      :unused => true, # Drop unreferenced functions and variables
+      :unused => true, # Drop unreferenced functions and variables (simple direct variable assignments do not count as references unless set to `"keep_assign"`)
+      :toplevel => false, # Drop unreferenced top-level functions and variables
+      :top_retain => [], # prevent specific toplevel functions and variables from `unused` removal
       :hoist_funs => true, # Hoist function declarations
       :hoist_vars => false, # Hoist var declarations
       :if_return => true, # Optimizations for if/return and if/continue