+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 - update UglifyJS to 2.8.0
 - add reduce_vars compress option
-- enable reduce_vars and collapse_vars compress options by default
+- enable `reduce_vars` and `collapse_vars` compress options by default
+- unused top-level function and variable removal with `toplevel` compress option
+- add `top_retain` compress option to specify list of top-level variables to always retain
 
 ## 3.0.4 (30 November 2016)